submittal_tracker/requirements.txt
playwright==1.49.1
pdfplumber==0.11.4
PyMuPDF==1.25.1
httpx==0.28.1
python-dotenv==1.0.1
gspread>=6.0.0
google-auth>=2.0.0
